Add documentation and configs
