+++ examples/tensorrt_llm/configs/llm_api_config_router.yaml
@@ -13,24 +13,26 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-Frontend:
-  served_model_name: deepseek-ai/DeepSeek-R1-Distill-Llama-8B
-  endpoint: dynamo.Processor.chat/completions
-  port: 8000
 
-Processor:
-  engine_args: "configs/llm_api_config_router.yaml"
-  block-size: 64
-  router: kv
+# In the case of disaggregated deployment, this config will apply to each server
+# and will be overwritten by the disaggregated config file
 
-Router:
-  model-name: deepseek-ai/DeepSeek-R1-Distill-Llama-8B
-  min-workers: 1
+model_name: "TinyLlama/TinyLlama-1.1B-Chat-v1.0"
+model_path: null
+tensor_parallel_size: 1
+moe_expert_parallel_size: 1
+enable_attention_dp: false
+max_num_tokens: 10240
+max_batch_size: 16
+trust_remote_code: true
+backend: pytorch
 
-TensorRTLLMWorker:
-  engine_args: "configs/llm_api_config_router.yaml"
-  router: kv
-  ServiceArgs:
-    workers: 1
-    resources:
-      gpu: 1
+kv_cache_config:
+  free_gpu_memory_fraction: 0.95
+  event_buffer_max_size: 1024
+  enable_block_reuse: true
+
+pytorch_backend_config:
+  enable_overlap_scheduler: false
+  use_cuda_graph: false
+  enable_iter_perf_stats: true
